+++ R/utils.R
@@ -661,6 +661,7 @@ set_scale_breaks <- function(plot_borders) {
 
 
 filter_years <- function(dat, years) {
+
   # if(is.null(years)){
   #   l <- consecutive_numbers(dat$year)
   # }
@@ -689,7 +690,6 @@ process_bundesland <- function(vec, linebreak = FALSE, total_group = "total"){
     vec <- gsub("([a-z])([A-Z])", "\\1\u2212 <br> \\2", vec)
   }else{
     vec <- gsub("([a-z])([A-Z])", "\\1\u2212\\2", vec)
-
 }
 
   vec <- gsub(total_group, "Deutschland", vec)