+++ R/utils_nudge.R
@@ -16,12 +16,12 @@ calc_brace_coords <- function(grouping_var_lvls, coords, years_list, plot_settin
   if (nrow(years_list$years_braces) == 0) {
     return(list(coord_dat = data.frame(), group_labels = data.frame(label_pos_y = 0)))
   }
-
   years_braces <- years_list$years_braces
 
 
   ## Calculate the y coordinates for the braces and labels:
   ## Check if any braces overlap. If that's the case, they have to plotted below each other
+
   overlap <- calc_overlap(
     year_start = years_braces$year_start,
     year_end = years_braces$year_end
@@ -32,7 +32,7 @@ calc_brace_coords <- function(grouping_var_lvls, coords, years_list, plot_settin
   # calc starting points for braces -----------------------------------------
   ## Braces should start some distance below the lower x-axis. Labels below lowest brace
   range_coords <- diff(range(coords))
-  if (overlap == TRUE) {
+  if (overlap) {
     lower_brace_y_a <- coords[1] - (range_coords * plot_settings$brace_span_y)
     lower_brace_y_b <- lower_brace_y_a - range_coords * plot_settings$brace_span_y
     upper_label_y <- lower_brace_y_b - range_coords * plot_settings$brace_label_nudge_y
@@ -96,12 +96,11 @@ calc_brace_coords <- function(grouping_var_lvls, coords, years_list, plot_settin
 calc_brace_position <- function(years_list, overlap) {
   ## Don't do this stuff df wise, only once for all year combinations.
   ## I only need one coordinate for each possible brace, that's it!
-
   years_braces_df <- years_list$years_braces
   ## Calculate the range between all year combinations:
-  years_braces_df$range <- apply(years_braces_df, 1, function(x) {
-    diff(range(x))
-  })
+
+  years_braces_df$range <- years_braces_df$year_end - years_braces_df$year_start
+
 
   if (overlap == TRUE) {
     if (all(years_braces_df$range == years_braces_df$range[1])) { ## All braces have the same range
@@ -119,6 +118,7 @@ calc_brace_position <- function(years_list, overlap) {
       )
     } else {
       # The smaller brace won't get any indention.
+
       years_braces_df$brace_position_x <- ifelse(years_braces_df$range == max(years_braces_df$range),
                                                  yes = ifelse(years_braces_df$year_start == min(years_braces_df$year_start),
                                                               yes = "left",
@@ -165,7 +165,6 @@ calc_brace_position <- function(years_list, overlap) {
 
 # Plot_points -------------------------------------------------------------
 calc_x_nudge <- function(dat, nudge_x, plot_settings) {
-
   x_value_range <- diff(dat$plot_lims$x_value_range)
   min_max_trend <- get_min_max(dat$dat_final)