+++ NEWS.md
@@ -1,6 +1,7 @@
 # eatPlot (development version)
 * `columns_nudge_y` in `plotsettings_tablebarplot()` now works rowwise as well. 
 * Improved rendering of greek letters with linebreak, using lists instead of ´atop()´.
+* Implemented more modular options for plotting lines for different years over different facets in `plot_lineplot()`. 
 
 # eatPlot 1.1.0
 * Internally changed the data preperation workflow.